apps/manager/requirements.txt
# Flask Web Framework
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-security-too==5.4.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Additional dependencies for Flask-Security-Too
email-validator==2.1.0
bcrypt==4.1.1

# Logging and monitoring
python-json-logger==2.0.7
